+++ atlasdb-config/src/main/java/com/palantir/atlasdb/factory/TransactionManagers.java
@@ -70,7 +70,9 @@ import com.palantir.atlasdb.persistentlock.KvsBackedPersistentLockService;
 import com.palantir.atlasdb.persistentlock.NoOpPersistentLockService;
 import com.palantir.atlasdb.persistentlock.PersistentLockService;
 import com.palantir.atlasdb.qos.QosClient;
+import com.palantir.atlasdb.qos.QosService;
 import com.palantir.atlasdb.qos.client.AtlasDbQosClient;
+import com.palantir.atlasdb.qos.config.ImmutableQosLimitsConfig;
 import com.palantir.atlasdb.qos.config.QosClientConfig;
 import com.palantir.atlasdb.qos.ratelimit.QosRateLimiters;
 import com.palantir.atlasdb.schema.generated.SweepTableFactory;
@@ -113,6 +115,9 @@ import com.palantir.lock.client.TimeLockClient;
 import com.palantir.lock.impl.LegacyTimelockService;
 import com.palantir.lock.impl.LockServiceImpl;
 import com.palantir.lock.v2.TimelockService;
+import com.palantir.remoting.api.config.service.ServiceConfiguration;
+import com.palantir.remoting3.clients.ClientConfigurations;
+import com.palantir.remoting3.jaxrs.JaxRsClient;
 import com.palantir.timestamp.TimestampService;
 import com.palantir.timestamp.TimestampStoreInvalidator;
 import com.palantir.tritium.metrics.registry.DefaultTaggedMetricRegistry;
@@ -170,6 +175,7 @@ public abstract class TransactionManagers {
 
     /**
      * Accepts a single {@link Schema}.
+     *
      * @see TransactionManagers#createInMemory(Set)
      */
     public static SerializableTransactionManager createInMemory(Schema schema) {
@@ -299,9 +305,26 @@ public abstract class TransactionManagers {
     }
 
     private QosClient getQosClient(Supplier<QosClientConfig> config) {
-        QosRateLimiters rateLimiters = QosRateLimiters.create(
-                JavaSuppliers.compose(conf -> conf.maxBackoffSleepTime().toMilliseconds(), config),
-                JavaSuppliers.compose(QosClientConfig::limits, config));
+        Optional<ServiceConfiguration> qosServiceConfig = config.get().qosService();
+        QosRateLimiters rateLimiters;
+        if (qosServiceConfig.isPresent()) {
+            QosService qosService = JaxRsClient.create(QosService.class, userAgent(),
+                    ClientConfigurations.of(qosServiceConfig.get()));
+            rateLimiters = QosRateLimiters.create(
+                    JavaSuppliers.compose(conf -> conf.maxBackoffSleepTime().toMilliseconds(), config),
+                    () -> {
+                        long readLimit = qosService.readLimit(config().getNamespaceString());
+                        long writeLimit = qosService.writeLimit(config().getNamespaceString());
+                        return ImmutableQosLimitsConfig.builder()
+                                .readBytesPerSecond(readLimit)
+                                .writeBytesPerSecond(writeLimit)
+                                .build();
+                    });
+        } else {
+            rateLimiters = QosRateLimiters.create(
+                    JavaSuppliers.compose(conf -> conf.maxBackoffSleepTime().toMilliseconds(), config),
+                    JavaSuppliers.compose(QosClientConfig::limits, config));
+        }
         return AtlasDbQosClient.create(rateLimiters);
     }