+++ qos-service-impl/src/integTest/java/com/palantir/atlasdb/qos/server/QosServerConfig.java
@@ -13,23 +13,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.palantir.atlasdb.qos.server;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.palantir.atlasdb.qos.config.QosServiceInstallConfig;
 import com.palantir.atlasdb.qos.config.QosServiceRuntimeConfig;
 
 import io.dropwizard.Configuration;
 
 public class QosServerConfig extends Configuration {
-
     private final QosServiceRuntimeConfig runtime;
+    private final QosServiceInstallConfig install;
 
-    public QosServerConfig(@JsonProperty("runtime") QosServiceRuntimeConfig runtime) {
+    public QosServerConfig(@JsonProperty("runtime") QosServiceRuntimeConfig runtime,
+            @JsonProperty("install") QosServiceInstallConfig install) {
         this.runtime = runtime;
+        this.install = install;
     }
 
     public QosServiceRuntimeConfig runtime() {
         return runtime;
+    }
+
+    public QosServiceInstallConfig install() {
+        return install;
     }
 }